+++ microhapulator/cli/pipe.py
@@ -196,6 +196,12 @@ def subparser(subparsers):
         # Hidden option for testing purposes
     )
     cli.add_argument(
+        "--hspace",
+        metavar="HS",
+        default=-0.7,
+        help="horizontal spacing between samples in the read distribution length ridge plots; negative value for this parameter enables overlapping plots; HS=-0.7 by default",
+    )
+    cli.add_argument(
         "markerrefr", help="path to a FASTA file containing marker reference sequences"
     )
     cli.add_argument("markerdefn", help="path to a TSV file containing marker definitions")
@@ -231,6 +237,7 @@ def main(args):
         thresh_file=args.config,
         paired=args.reads_are_paired,
         ambiguous_thresh=args.ambiguous_thresh,
+        hspace=args.hspace,
     )
     snakefile = resource_filename("microhapulator", "workflows/analysis.smk")
     success = snakemake(