+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Smart handling of multiple markers at a single locus implemented for `type` and `pipe` (#158).
 - Replaced `parsers` module with a `MicrohapIndex` class and supporting classes (#158).
+- Replaced read length distubition histograms with ridge plots (#167)
 
 ### Fixed
 - Bug with inconsistent sorting of read counts for interlocus balance (#159).